Refactor, original scheme
